fix(nontransparent): move auto-detect code out of Ragel-generated file
nontransparent/parser.go is regenerated from parser.go.rl by Ragel.
The auto-detect code (NewParserAuto, SetRFC3164Options, etc.) was added
directly to the generated file, so make + gofmt on Go 1.24 wiped it,
causing the git-diff-exit-code CI check to fail.

Move the auto-detect functions and tests to separate hand-written files
(parser_auto.go, parser_auto_test.go) and add the struct fields to the
.rl source so they survive regeneration.
